+++ modules/math/externals/Shoemake/EulerAngles.cpp
@@ -17,7 +17,7 @@ Quat Eul_ToQuat(EulerAngles ea)
     Quat qu;
     double a[3], ti, tj, th, ci, cj, ch, si, sj, sh, cc, cs, sc, ss;
     int i,j,k,/*h,*/n,s,f;
-    EulGetOrd(ea.w,i,j,k,/*h,*/n,s,f);
+    EulGetOrd((int)ea.w,i,j,k,/*h,*/n,s,f);
     if (f==EulFrmR) {float t = ea.x; ea.x = ea.z; ea.z = t;}
     if (n==EulParOdd) ea.y = -ea.y;
     ti = ea.x*0.5; tj = ea.y*0.5; th = ea.z*0.5;
@@ -29,16 +29,18 @@ Quat Eul_ToQuat(EulerAngles ea)
         a[i] = cj*(cs + sc);	/* Could speed up with */
         a[j] = sj*(cc + ss);	/* trig identities. */
         a[k] = sj*(cs - sc);
-        qu.w = cj*(cc - ss);
+        qu.w = (float)(cj*(cc - ss));
     } else
     {
         a[i] = cj*sc - sj*cs;
         a[j] = cj*ss + sj*cc;
         a[k] = cj*cs - sj*sc;
-        qu.w = cj*cc + sj*ss;
+        qu.w = (float)(cj * cc + sj * ss);
     }
     if (n==EulParOdd) a[j] = -a[j];
-    qu.x = a[X]; qu.y = a[Y]; qu.z = a[Z];
+    qu.x = (float)a[X]; 
+    qu.y = (float)a[Y]; 
+    qu.z = (float)a[Z];
     return (qu);
 }
 
@@ -46,7 +48,7 @@ Quat Eul_ToQuat(EulerAngles ea)
 void Eul_ToHMatrix(EulerAngles ea, HMatrix M)
 {  double ti, tj, th, ci, cj, ch, si, sj, sh, cc, cs, sc, ss;
    int i,j,k,/*h,*/n,s,f;
-   EulGetOrd(ea.w,i,j,k,/*h,*/n,s,f);
+   EulGetOrd((int)ea.w,i,j,k,/*h,*/n,s,f);
    if (f==EulFrmR) {float t = ea.x; ea.x = ea.z; ea.z = t;}
    if (n==EulParOdd) {ea.x = -ea.x; ea.y = -ea.y; ea.z = -ea.z;}
    ti = ea.x;	  tj = ea.y;	th = ea.z;
@@ -74,29 +76,29 @@ EulerAngles Eul_FromHMatrix(HMatrix M, int order)
     if (s==EulRepYes) {
    double sy = sqrt(M[i][j]*M[i][j] + M[i][k]*M[i][k]);
    if (sy > 16*FLT_EPSILON) {
-       ea.x = atan2(M[i][j], M[i][k]);
-       ea.y = atan2((float)sy, M[i][i]);
-       ea.z = atan2(M[j][i], -M[k][i]);
+       ea.x = (float)atan2(M[i][j], M[i][k]);
+       ea.y = (float)atan2((float)sy, M[i][i]);
+       ea.z = (float)atan2(M[j][i], -M[k][i]);
    } else {
-       ea.x = atan2(-M[j][k], M[j][j]);
-       ea.y = atan2((float)sy, M[i][i]);
-       ea.z = 0;
+       ea.x = (float)atan2(-M[j][k], M[j][j]);
+       ea.y = (float)atan2((float)sy, M[i][i]);
+       ea.z = 0.0f;
    }
     } else {
    double cy = sqrt(M[i][i]*M[i][i] + M[j][i]*M[j][i]);
    if (cy > 16*FLT_EPSILON) {
-       ea.x = atan2(M[k][j], M[k][k]);
-       ea.y = atan2(-M[k][i], (float)cy);
-       ea.z = atan2(M[j][i], M[i][i]);
+       ea.x = (float)atan2(M[k][j], M[k][k]);
+       ea.y = (float)atan2(-M[k][i], (float)cy);
+       ea.z = (float)atan2(M[j][i], M[i][i]);
    } else {
-       ea.x = atan2(-M[j][k], M[j][j]);
-       ea.y = atan2(-M[k][i], (float)cy);
-       ea.z = 0;
+       ea.x = (float)atan2(-M[j][k], M[j][j]);
+       ea.y = (float)atan2(-M[k][i], (float)cy);
+       ea.z = 0.0f;
    }
     }
     if (n==EulParOdd) {ea.x = -ea.x; ea.y = - ea.y; ea.z = -ea.z;}
     if (f==EulFrmR) {float t = ea.x; ea.x = ea.z; ea.z = t;}
-    ea.w = order;
+    ea.w = (float)order;
     return (ea);
 }
